refactor : controller 에서 관리되던 핵심 기능을 service 에서 담당하도록 변경
