+++ src/test/kotlin/service/RaceControllerTest.kt
@@ -1,11 +1,12 @@
-package controller
+package service
 
+import controller.OutController
 import model.Car
-import org.assertj.core.api.Assertions.assertThat
+import model.RaceService
+import org.assertj.core.api.Assertions
 import org.junit.jupiter.api.AfterEach
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
-import view.InView
 import view.OutView
 import java.io.ByteArrayOutputStream
 import java.io.OutputStream
@@ -35,10 +36,10 @@ class RaceControllerTest {
 
     @Test
     fun `랜덤으로 움직인 자동차 중 가장 많이 움직인 자동차가 우승한다`() {
-        val raceController = RaceController(4, cars, outController)
-        raceController.fullRace()
-        val winnerList = raceController.getFinalResult()
+        val raceService = RaceService(4, cars, outController)
+        raceService.fullRace()
+        val winnerList = raceService.getFinalResult()
         outController.printFinalResult(winnerList)
-        assertThat(output()).contains("최종 우승자 : ${winnerList.joinToString(", ") { it.name }}")
+        Assertions.assertThat(output()).contains("최종 우승자 : ${winnerList.joinToString(", ") { it.name }}")
     }
 }
